Prevent (sign-)conversion warning
See also gsl-lite PR 71, https://github.com/martinmoene/gsl-lite/pull/71

+++ include/lest/lest.hpp
@@ -393,7 +393,7 @@ public:
     friend bool operator != ( approx const & lhs, double rhs ) { return !operator==( rhs, lhs ); }
 
     friend bool operator <= ( double lhs, approx const & rhs ) { return lhs < rhs.magnitude_ || lhs == rhs; }
-    friend bool operator <= ( approx const & lhs, double rhs ) { return lhs.magnitude_ < rhs || lhs == rhs; } 
+    friend bool operator <= ( approx const & lhs, double rhs ) { return lhs.magnitude_ < rhs || lhs == rhs; }
     friend bool operator >= ( double lhs, approx const & rhs ) { return lhs > rhs.magnitude_ || lhs == rhs; }
     friend bool operator >= ( approx const & lhs, double rhs ) { return lhs.magnitude_ > rhs || lhs == rhs; }
 
@@ -1090,7 +1090,7 @@ inline void shuffle( tests & specification, options option )
 
 inline int stoi( text num )
 {
-    return std::strtol( num.c_str(), NULL, 10 );
+    return static_cast<int>( std::strtol( num.c_str(), NULL, 10 ) );
 }
 
 inline bool is_number( text arg )
